+++ addons/coupon/data/coupon_email_data.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" ?>
 <odoo>
    <data noupdate="1">
-      <record id="sale_coupon.mail_template_sale_coupon" model="mail.template">
+      <record id="coupon.mail_template_sale_coupon" model="mail.template">
          <field name="name">Coupon: Send by Email</field>
-         <field name="model_id" ref="sale_coupon.model_sale_coupon"/>
+         <field name="model_id" ref="coupon.model_coupon_coupon"/>
          <field name="subject">Your reward coupon from ${object.program_id.company_id.name} </field>
          <field name="email_from">${object.program_id.company_id.email | safe}</field>
          <field name="partner_to">${object.partner_id.id}</field>
@@ -103,7 +103,7 @@
 
         <record id="expire_coupon_cron" model="ir.cron">
             <field name="name">Coupon: expire coupon based on date</field>
-            <field name="model_id" ref="sale_coupon.model_sale_coupon"/>
+            <field name="model_id" ref="coupon.model_coupon_coupon"/>
             <field name="state">code</field>
             <field name="code">model.cron_expire_coupon()</field>
             <field name="active" eval="True"/>
