+++ addons/coupon/demo/coupon_demo.xml
@@ -8,12 +8,11 @@
             <field name="supplier_taxes_id" eval="False"/>
             <field name="sale_ok" eval="False"/>
             <field name="purchase_ok" eval="False"/>
-            <field name="invoice_policy">order</field>
             <field name="default_code">10PERCENTDISC</field>
             <field name="categ_id" ref="product.product_category_3"/>
         </record>
 
-        <record id="10_percent_auto_applied" model="sale.coupon.program">
+        <record id="10_percent_auto_applied" model="coupon.program">
             <field name="name">Code for 10% on orders</field>
             <field name="promo_code_usage">code_needed</field>
             <field name="promo_code">10pc</field>
@@ -21,7 +20,7 @@
             <field name="discount_type">percentage</field>
             <field name="discount_percentage">10.0</field>
             <field name="program_type">promotion_program</field>
-            <field name="discount_line_product_id" ref="sale_coupon.product_product_10_percent_discount"/>
+            <field name="discount_line_product_id" ref="coupon.product_product_10_percent_discount"/>
             <field name="validity_duration">0</field>
         </record>
 
@@ -32,12 +31,11 @@
             <field name="supplier_taxes_id" eval="False"/>
             <field name="sale_ok" eval="False"/>
             <field name="purchase_ok" eval="False"/>
-            <field name="invoice_policy">order</field>
             <field name="default_code">FREELARGECABINET</field>
             <field name="categ_id" ref="product.product_category_3"/>
         </record>
 
-        <record id="3_cabinets_plus_1_free" model="sale.coupon.program">
+        <record id="3_cabinets_plus_1_free" model="coupon.program">
             <field name="name">Buy 3 large cabinets, get one for free</field>
             <field name="promo_code_usage">no_code_needed</field>
             <field name="discount_apply_on">on_order</field>
@@ -46,11 +44,11 @@
             <field name="reward_product_id" ref="product.product_product_6"></field>
             <field name="rule_min_quantity">3</field>
             <field name="rule_products_domain">[["name","ilike","large cabinet"]]</field>
-            <field name="discount_line_product_id" ref="sale_coupon.product_product_free_large_cabinet"/>
+            <field name="discount_line_product_id" ref="coupon.product_product_free_large_cabinet"/>
             <field name="validity_duration">0</field>
         </record>
 
-        <record id="10_percent_coupon" model="sale.coupon.program">
+        <record id="10_percent_coupon" model="coupon.program">
             <field name="name">10% Discount</field>
             <field name="promo_code_usage">code_needed</field>
             <field name="discount_apply_on">on_order</field>
