+++ addons/coupon/models/__init__.py
@@ -2,8 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import mail_compose_message
-from . import sale_coupon
-from . import sale_coupon_reward
-from . import sale_coupon_rules
-from . import sale_coupon_program
-from . import sale_order
+from . import coupon
+from . import coupon_reward
+from . import coupon_rules
+from . import coupon_program
