+++ addons/coupon/report/coupon_report.xml
@@ -3,9 +3,9 @@
     <data>
         <report
             id="report_coupon_code" 
-            model="sale.coupon"
-            name="sale_coupon.report_coupon_i18n"
-            file="sale_coupon.report_coupon_i18n"
+            model="coupon.coupon"
+            name="coupon.report_coupon_i18n"
+            file="coupon.report_coupon_i18n"
             report_type="qweb-pdf"
             string="Coupon Code"/>
     </data>
