+++ addons/coupon/report/coupon_report_templates.xml
@@ -94,7 +94,7 @@
         <t t-call="web.html_container">
             <t t-foreach="docs" t-as="o">
                 <t t-set="o" t-value="o.with_context(lang=o.partner_id.lang)"/>
-                <t t-call="sale_coupon.report_coupon" t-lang="o.partner_id.lang"/>
+                <t t-call="coupon.report_coupon" t-lang="o.partner_id.lang"/>
             </t>
         </t>
     </template>
