+++ addons/coupon/security/coupon_security.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+    <record id="sale_coupon_generate_rule" model="ir.rule">
+        <field name="name">Generate Sales Coupon Rule</field>
+        <field name="model_id" ref="model_coupon_generate_wizard"/>
+        <field name="domain_force">[('create_uid', '=', user.id)]</field>
+    </record>
+</odoo>
