+++ addons/coupon/wizard/coupon_generate.py
@@ -6,9 +6,9 @@ import ast
 from odoo.osv import expression
 
 
-class SaleCouponGenerate(models.TransientModel):
-    _name = 'sale.coupon.generate'
-    _description = 'Generate Sales Coupon'
+class CouponGenerate(models.TransientModel):
+    _name = 'coupon.generate.wizard'
+    _description = 'Generate Coupon'
 
     nbr_coupons = fields.Integer(string="Number of Coupons", help="Number of coupons", default=1)
     generation_type = fields.Selection([
@@ -21,20 +21,20 @@ class SaleCouponGenerate(models.TransientModel):
     def generate_coupon(self):
         """Generates the number of coupons entered in wizard field nbr_coupons
         """
-        program = self.env['sale.coupon.program'].browse(self.env.context.get('active_id'))
+        program = self.env['coupon.program'].browse(self.env.context.get('active_id'))
 
         vals = {'program_id': program.id}
 
         if self.generation_type == 'nbr_coupon' and self.nbr_coupons > 0:
             for count in range(0, self.nbr_coupons):
-                self.env['sale.coupon'].create(vals)
+                self.env['coupon.coupon'].create(vals)
 
         if self.generation_type == 'nbr_customer' and self.partners_domain:
             for partner in self.env['res.partner'].search(ast.literal_eval(self.partners_domain)):
                 vals.update({'partner_id': partner.id, 'state': 'sent' if partner.email else 'new'})
-                coupon = self.env['sale.coupon'].create(vals)
+                coupon = self.env['coupon.coupon'].create(vals)
                 subject = '%s, a coupon has been generated for you' % (partner.name)
-                template = self.env.ref('sale_coupon.mail_template_sale_coupon', raise_if_not_found=False)
+                template = self.env.ref('coupon.mail_template_sale_coupon', raise_if_not_found=False)
                 if template:
                     email_values = {'email_to': partner.email, 'email_from': self.env.user.email or '', 'subject': subject}
                     template.send_mail(coupon.id, email_values=email_values, notif_layout='mail.mail_notification_light')
