+++ addons/coupon/wizard/coupon_generate_views.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="sale_coupon_generate_view_form" model="ir.ui.view">
-        <field name="name">sale.coupon.generate.form</field>
-        <field name="model">sale.coupon.generate</field>
+    <record id="coupon_generate_view_form" model="ir.ui.view">
+        <field name="name">coupon.generate.wizard.form</field>
+        <field name="model">coupon.generate.wizard</field>
         <field name="arch" type="xml">
             <form string="Generate Coupons">
                 <group>
@@ -22,11 +22,11 @@
         </field>
     </record>
 
-    <record id="sale_coupon_generate_action" model="ir.actions.act_window">
+    <record id="coupon_generate_action" model="ir.actions.act_window">
         <field name="name">Number of Coupons To Generate</field>
-        <field name="res_model">sale.coupon.generate</field>
+        <field name="res_model">coupon.generate.wizard</field>
         <field name="view_mode">form</field>
         <field name="target">new</field>
-        <field name="view_id" ref="sale_coupon_generate_view_form"/>
+        <field name="view_id" ref="coupon_generate_view_form"/>
     </record>
 </odoo>
