+++ addons/sale_coupon/__init__.py
@@ -3,4 +3,3 @@
 
 from . import models
 from . import wizard
-from . import report
