+++ addons/sale_coupon/models/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import coupon_generate
+from . import coupon
+from . import coupon_program
+from . import sale_order
