+++ addons/sale_coupon_delivery/models/sale_coupon_program.py
@@ -5,14 +5,14 @@
 from odoo import models, _, api
 
 
-class SaleCouponProgram(models.Model):
-    _inherit = "sale.coupon.program"
+class CouponProgram(models.Model):
+    _inherit = "coupon.program"
 
     def _filter_not_ordered_reward_programs(self, order):
         """
         Returns the programs when the reward is actually in the order lines
         """
-        programs = super(SaleCouponProgram, self)._filter_not_ordered_reward_programs(order)
+        programs = super(CouponProgram, self)._filter_not_ordered_reward_programs(order)
         # Do not filter on free delivery programs. As delivery_unset is called everywhere (which is
         # rather stupid), the delivery line is unliked to be created again instead of writing on it to
         # modify the price_unit. That way, the reward is unlink and is not set back again.
@@ -21,4 +21,4 @@ class SaleCouponProgram(models.Model):
     def _check_promo_code(self, order, coupon_code):
         if self.reward_type == 'free_shipping' and not order.order_line.filtered(lambda line: line.is_delivery):
             return {'error': _('The shipping costs are not in the order lines.')}
-        return super(SaleCouponProgram, self)._check_promo_code(order, coupon_code)
+        return super(CouponProgram, self)._check_promo_code(order, coupon_code)
