+++ addons/sale_coupon_delivery/models/sale_order.py
@@ -9,13 +9,13 @@ class SaleOrder(models.Model):
     def _get_no_effect_on_threshold_lines(self):
         self.ensure_one()
         # Do not count shipping and free shipping
-        free_delivery_product = self.env['sale.coupon.program'].search([('reward_type', '=', 'free_shipping')]).mapped('discount_line_product_id')
+        free_delivery_product = self.env['coupon.program'].search([('reward_type', '=', 'free_shipping')]).mapped('discount_line_product_id')
         lines = self.order_line.filtered(lambda line: line.is_delivery or line.product_id in free_delivery_product)
         return lines + super(SaleOrder, self)._get_no_effect_on_threshold_lines()
 
     def _get_paid_order_lines(self):
         """ Returns the taxes included sale order total amount without the rewards amount"""
-        free_reward_product = self.env['sale.coupon.program'].search([('reward_type', '=', 'product')]).mapped('discount_line_product_id')
+        free_reward_product = self.env['coupon.program'].search([('reward_type', '=', 'product')]).mapped('discount_line_product_id')
         return self.order_line.filtered(lambda x: not (x.is_reward_line or x.is_delivery) or x.product_id in free_reward_product)
 
     def _get_reward_line_values(self, program):
