+++ addons/website_sale_coupon/models/sale_coupon_program.py
@@ -5,9 +5,9 @@ from odoo import api, models, _
 from odoo.exceptions import ValidationError
 
 
-class SaleCouponProgram(models.Model):
-    _name = 'sale.coupon.program'
-    _inherit = ['sale.coupon.program', 'website.multi.mixin']
+class CouponProgram(models.Model):
+    _name = 'coupon.program'
+    _inherit = ['coupon.program', 'website.multi.mixin']
 
     @api.constrains('promo_code', 'website_id')
     def _check_promo_code_constraint(self):
