+++ addons/website_sale_coupon/views/sale_coupon_program_views.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <menuitem action="sale_coupon.sale_coupon_program_action_promo_program" id="menu_promotion_type_config" name="Promotion Programs" parent="website_sale.menu_catalog" groups="sales_team.group_sale_manager" sequence="50"/>
-    <menuitem id="menu_coupon_type_config" action="sale_coupon.sale_coupon_program_action_coupon_program" name="Coupon Programs" parent="website_sale.menu_catalog" groups="sales_team.group_sale_manager" sequence="51"/>
+    <menuitem action="coupon.coupon_program_action_promo_program" id="menu_promotion_type_config" name="Promotion Programs" parent="website_sale.menu_catalog" groups="sales_team.group_sale_manager" sequence="50"/>
+    <menuitem id="menu_coupon_type_config" action="coupon.coupon_program_action_coupon_program" name="Coupon Programs" parent="website_sale.menu_catalog" groups="sales_team.group_sale_manager" sequence="51"/>
 
     <record model="ir.ui.view" id="sale_coupon_program_view_form_common_website">
-        <field name="name">sale.coupon.program.common.form</field>
-        <field name="model">sale.coupon.program</field>
-        <field name="inherit_id" ref="sale_coupon.sale_coupon_program_view_form_common"/>
+        <field name="name">coupon.program.common.form</field>
+        <field name="model">coupon.program</field>
+        <field name="inherit_id" ref="coupon.coupon_program_view_form_common"/>
         <field name="arch" type="xml">
             <group name="validity" position="inside">
                 <label for="website_id" groups="website.group_multi_website"/>
@@ -19,9 +19,9 @@
     </record>
 
     <record id="sale_coupon_program_view_tree_website" model="ir.ui.view">
-        <field name="name">sale.coupon.program.tree</field>
-        <field name="model">sale.coupon.program</field>
-        <field name="inherit_id" ref="sale_coupon.sale_coupon_program_view_tree"/>
+        <field name="name">coupon.program.tree</field>
+        <field name="model">coupon.program</field>
+        <field name="inherit_id" ref="coupon.coupon_program_view_tree"/>
         <field name="arch" type="xml">
             <field name="name" position="after">
                 <field name="website_id" groups="website.group_multi_website"/>
